Conflicts:
	Data/Dialogs/dlgConfiguration.xml
	Data/Dialogs/dlgConfiguration_L.xml
	doc/manual/en/ch11_configuration.tex

+++ src/LocalTime.hpp
@@ -29,6 +29,8 @@ int DetectCurrentTime(const NMEA_INFO *Basic);
 
 int TimeLocal(int d, int utc_offset);
 int TimeLocal(int d);
-long GetUTCOffset(void);
+
+int
+GetUTCOffset(void);
 
 #endif
